Clean up the main file again
